added name validation for team

+++ src/main/java/edu/sjsu/cmpe275/lab2/phoenix/service/impl/TeamServiceImpl.java
@@ -113,6 +113,8 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public Team deleteTeam(String id){
         Team team = teamRepository.findById(id).orElse(null);
+        if(team==null)
+            throw new TeamNotFoundException();
         team.getPlayers().forEach(player -> player.setTeam(null));
         playerRepository.saveAll(team.getPlayers());
         teamRepository.deleteById(id);
